requirements.txt
soundfile>=0.12.0
numpy>=1.21.0
pytest>=7.0.0
allure-pytest>=2.13.0
